+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to
   sync with the number of search results actually being displayed. (#827)
 * Hide app directory search loading indicator when search returns zero results
   (#826)
+* Suggest ways to recover from a search with zero results (#828)
 * Gracefully handle case where directory search JSON URL is bad, as is the case
   in naive localhost demo against stub data (#825)